requirements.txt
alembic==1.15.2
aniso8601==10.0.1
bcrypt==4.3.0
blinker==1.9.0
click==8.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-restful-swagger==0.20.2
flask-restful-swagger-3==0.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
